+++ lib/chef/application/client.rb
@@ -18,8 +18,6 @@
 # limitations under the License.
 
 require_relative "base"
-require_relative "../handler/error_report"
-require_relative "../workstation_config_loader"
 require "uri" unless defined?(URI)
 
 # DO NOT MAKE EDITS, see Chef::Application::Base
@@ -31,22 +29,8 @@ class Chef::Application::Client < Chef::Application::Base
   option :config_file,
     short: "-c CONFIG",
     long: "--config CONFIG",
+        default: Chef::Config.platform_specific_path("#{Chef::Dist::CONF_DIR}/solo.rb"),
     description: "The configuration file to use."
-
-  unless Chef::Platform.windows?
-    option :daemonize,
-      short: "-d [WAIT]",
-      long: "--daemonize [WAIT]",
-      description: "Daemonize the process. Accepts an optional integer which is the " \
-        "number of seconds to wait before the first daemonized run.",
-      proc: lambda { |wait| wait =~ /^\d+$/ ? wait.to_i : true }
-  end
-
-  option :pid_file,
-    short: "-P PID_FILE",
-    long: "--pid PIDFILE",
-    description: "Set the PID file location, for the #{Chef::Dist::CLIENT} daemon process. Defaults to /tmp/chef-client.pid.",
-    proc: nil
 
   option :runlist,
     short: "-r RunlistItem,RunlistItem...",
@@ -67,25 +51,4 @@ class Chef::Application::Client < Chef::Application::Base
     @solo_flag = solo
     super()
   end
-
-  def run(enforce_license: false)
-    setup_signal_handlers
-    reconfigure
-    # setup_application does a Dir.chdir("/") and cannot come before reconfigure or many things break
-    setup_application
-    check_license_acceptance if enforce_license
-    for_ezra if Chef::Config[:ez]
-    if Chef::Config[:solo_legacy_mode]
-      Chef::Application::Solo.new.run # FIXME: minimally we just need to reparse the cli and then run_application
-    else
-      run_application
-    end
-  end
-
-  def configure_logging
-    super
-    Mixlib::Authentication::Log.use_log_devices( Chef::Log )
-    Ohai::Log.use_log_devices( Chef::Log )
-  end
-
 end
